completing todo form component
