+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "vite-project",
+  "name": "-react-todo",
   "private": true,
   "version": "0.0.0",
   "type": "module",
