+++ vite-project/src/App.jsx
@@ -1,5 +1,4 @@
 
-export default App
 import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
